Rename lib_sdk folder to lib
