+++ api_tests/dry/rfc003_swap_reject.ts
@@ -1,14 +1,14 @@
 // These are stateless tests -- they don't require any state of the cnd and they don't change it
 // They are mostly about checking invalid request responses
 // These test do not use the sdk so that we can test edge cases
-import { twoActorTest } from "../lib_sdk/actor_test";
+import { twoActorTest } from "../lib/actor_test";
 import { expect, request } from "chai";
 import "chai/register-should";
-import "../lib_sdk/setup_chai";
+import "../lib/setup_chai";
 import { EmbeddedRepresentationSubEntity } from "../gen/siren";
 import * as swapPropertiesJsonSchema from "../swap.schema.json";
-import { Actor } from "../lib_sdk/actors/actor";
-import { createDefaultSwapRequest, DEFAULT_ALPHA } from "../lib_sdk/utils";
+import { Actor } from "../lib/actors/actor";
+import { createDefaultSwapRequest, DEFAULT_ALPHA } from "../lib/utils";
 
 async function assertSwapsInProgress(actor: Actor, message: string) {
     const res = await request(actor.cndHttpApiUrl()).get("/swaps");
