+++ api_tests/e2e/rfc003/btc_eth/resume.ts
@@ -1,6 +1,6 @@
-import { twoActorTest } from "../../../lib_sdk/actor_test";
-import { AssetKind } from "../../../lib_sdk/asset";
-import { sleep } from "../../../lib_sdk/utils";
+import { twoActorTest } from "../../../lib/actor_test";
+import { AssetKind } from "../../../lib/asset";
+import { sleep } from "../../../lib/utils";
 
 setTimeout(function() {
     twoActorTest("rfc003-btc-eth-resume-alice-down-bob-funds", async function({
