+++ api_tests/e2e/rfc003/eth_btc/bitcoin_high_fee.ts
@@ -1,5 +1,5 @@
-import { twoActorTest } from "../../../lib_sdk/actor_test";
-import { AssetKind } from "../../../lib_sdk/asset";
+import { twoActorTest } from "../../../lib/actor_test";
+import { AssetKind } from "../../../lib/asset";
 import { expect } from "chai";
 
 setTimeout(function() {
