+++ api_tests/e2e/rfc003/eth_btc/ignore_fail_eth_transactions.ts
@@ -1,5 +1,5 @@
-import { twoActorTest } from "../../../lib_sdk/actor_test";
-import { AssetKind } from "../../../lib_sdk/asset";
+import { twoActorTest } from "../../../lib/actor_test";
+import { AssetKind } from "../../../lib/asset";
 
 setTimeout(function() {
     twoActorTest("rfc003-eth-btc-alpha-deploy-fails", async function({
