+++ api_tests/lib/wallets/bitcoin.ts
