+++ api_tests/lib/wallets/satoshi_bitcoin.d.ts
